changes some theming aspects to accept scss styling
removed some code lines that were either not meant to be there or that prevented some theming to occur that allowed it to work well with the scss styling and theming. Also, imported the UMN palette so that it can be used later if that is the route that we decide to go down.