Change to the positions of the graphs, room filter, and the the Select Room button #12 #8 #7
alters the positioning of the graphs to the bottom of the page. The room filter was removed and will hopefully be connected to the select room button later. the select room button will be opening a dialog where the user will see all the available rooms and their data (vacant, not vacant).