fix issue where text and icons where not displaying correct colors #7 #8 #6
Title text was not displaying in white text the entire time; dark mode would cause it to be black text. Icons for machines now are green for available, yellow for in use, and no color for broken (white or black based on theme). Progress bars now coordinate with machine colors.