remove menu sidenav and add to toolbar #7 #8 #6
removed the sidenav because it only had the home. Since there was not much change off of the page, we commented it out for now to see if there is anything that it could be used for later. Added a about and a Contact button that will be furnished later, but for now it was merely a test dummy to make sure we could get more things on the toolbar.