fix issue where notification dialog issue #15 #16
Radio buttons were not functioning properly. They are now fixed after some changes to the html for the sub dialog. Also, UI changes to the dialog and the rest of the site to reduce the border radius to fit more box style format. Sub button has some issues still, will be looking into later, which seems to be the only current viewable issue so far from the process.